requirements.txt
Django==4.2.3
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.17
python-dotenv==1.0.0

